added PersonDAO and interface. Adjusted GUI so that notable people are shown for batman begins

+++ src/controller/EditMovieController.java
@@ -1,19 +1,22 @@
 package controller;
 
-import data.DatabaseAccessObject;
+import data.MovieDAO;
+import data.PersonDAO;
+import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.TextField;
-import javafx.scene.layout.GridPane;
-import javafx.stage.Stage;
+import javafx.scene.control.*;
+import javafx.scene.image.ImageView;
 import model.Movie;
+import model.Person;
 
+import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Optional;
 import java.util.ResourceBundle;
-import java.util.UUID;
 
 public class EditMovieController implements Initializable {
 
@@ -28,10 +31,20 @@ public class EditMovieController implements Initializable {
     @FXML
     Button deleteButton;
     @FXML
-    GridPane gridPane;
+    TabPane tabPane;
+    @FXML
+    ImageView poster;
+    @FXML
+    Label labelCast;
+    @FXML
+    Label labelDirector;
+    @FXML
+    Label labelWriter;
 
     Movie movie;
-    DatabaseAccessObject dao = new DatabaseAccessObject();
+    ArrayList<Person> cast;
+    MovieDAO dao = new MovieDAO();
+
 
     public void saveMovieChanges(ActionEvent actionEvent){
         dao.connect();
@@ -47,18 +60,56 @@ public class EditMovieController implements Initializable {
     }
 
     public void deleteMovie(ActionEvent actionEvent){
-        dao.connect();
-        try {
-            movie.setTitle(insertTitleField.getText());
-            movie.setYear(Integer.parseInt(insertYearField.getText()));
-            movie.setPoster(insertPosterField.getText());
-        } catch (NumberFormatException | IOException e) {
-            System.out.println("year field only takes whole numbers");
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation Dialog");
+        alert.setHeaderText("This will permanently delete the movie from the database");
+        alert.setContentText("Are you ok with this?");
+
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK){
+            dao.connect();
+            try {
+                movie.setTitle(insertTitleField.getText());
+                movie.setYear(Integer.parseInt(insertYearField.getText()));
+                movie.setPoster(insertPosterField.getText());
+            } catch (NumberFormatException | IOException e) {
+                System.out.println("year field only takes whole numbers");
+            }
+            dao.deleteMovie(movie);
+            tabPane.getTabs().remove( tabPane.getSelectionModel().getSelectedIndex() );
+            tabPane.getSelectionModel()
+                    .selectFirst();
+        } else {
+            return;
         }
-        dao.deleteMovie(movie);
+
+
+    }
+
+    public void setLabel(Iterable<Person> cast){
+        StringBuilder sb = new StringBuilder();
+        for (Person p :
+                cast) {
+            if (p.getProfession().equals("Director")){
+                labelDirector.setText(p.toString());
+            } else if (p.getProfession().equals("Writer")){
+                labelWriter.setText(p.toString());
+            } else if (p.getProfession().equals("Actor")){
+                labelCast.setText(p.toString());
+            }
+            sb.append(p.toString()+"\n");
+        }
+        //labelCast.setText("Cast: "+sb);
+    }
+
+    public void setMovie(Movie movie){
+        insertTitleField.setText(movie.getTitle());
+        insertYearField.setText(Integer.toString(movie.getYear()));
+        insertPosterField.setText(movie.getPosterUrl());
+        poster.setImage(SwingFXUtils.toFXImage((BufferedImage) movie.getPoster(), null));
     }
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-
     }
 }
